+++ Store.Application/service/IProductService.cs
@@ -4,7 +4,7 @@ namespace Store.Application.service;
 
 public interface IProductService
 {
-    Task<CachedProduct> ChangePrice (Guid cachedProductId, float newPrice);
+    Task<CachedProduct> ChangePrice (Guid cachedProductId, double newPrice);
     Task<List<CachedProduct>> ShowAllProductsAsync();
     Task<CachedProduct> ShowProductDetails(string barcode );
 }
